Added hook for managing state in IO Select component (#510)
* Select state management improvement

Added a hook for handling the select state logic. The problem was that when only 1 option was available, the select didn't trigger as the setValue function was only triggered onChange... and you can't change a select with only 1 option.

* Select state management improvement

Added a hook for handling the select state logic. The problem was that when only 1 option was available, the select didn't trigger as the setValue function was only triggered onChange... and you can't change a select with only 1 option.

* moved hook

* Update useSelect.tsx

Added check to useSelect hook to not call the setting function when the value hasn't changed.

* Better use of hook for updating value

1. Added "No Wire" option to the List wire property.
2. Made sure the wire is not set if the new value is the same as current wire value.
3. If there is only 1 option available, set it.

* simplifying select component for wires

1. Removed useSelect hook
2. added noValueLabel prop to definitionSelectorProp

* Update conditionprop.tsx

* Update conditionprop.tsx