Inline styles builder #y1duf7 (#533)
* variants in components

* removed hardcoded grouping

* Added function to parse the variant namespace string

* Added logic to variant select fields. Also removed some repetitive code

* Added support for variant grouping

Refactored grouping logic into a function. With a new exception for the variant logic, the grouping code got a little chunky. A function with return statements seemed more appropriate to handle the extra logic

* Keyvaluelist component

Added keyvalue list component. Still placed in the wrong part of the project for now.

Classes key is  added to the component definition.

* definition state and styles

added some styling and getting def from app state

* Moved KeyValueList component

* empty label was rendered empty, causing a pixel bump

* Styling and optimization

* Fix build

* fix prettier

* reuse listfield component

* remove wrapper class

* fix autoadd

* small fix

Co-authored-by: Abel <abel@thecloudmasters.com>
Co-authored-by: humandad <55447225+humandad@users.noreply.github.com>