Bugfix and a bit of refactoring
Bug: when adding the first field in a new wire definition, the definition was printed out wrongly, making the page crash.

Reason: The 'fields' property is not in the definition when creating a new wire.

Fix: add a check to ensure the fields property is in the definition before adding fields.

Also, I've separated the logic a bit from the render function making the easier to read... but yeah, that's my humble opinion :)