bugfix
Case: user changes the component name to something unknown e.g. "i.box"
What happens: The properties panel tries to look up the properties for the path with that faulty component name... the app explodes :)

Maybe we need a better fix but this will keep the app from failing down.