Errorboundary component (#1190)
* Update slot.tsx

added check for array time. This keeps the map from failing bringing down the app when the listDef is not an array. This happens when the yaml is malformed.

* codepanel bugfixes

* Update slot.tsx

Filter out falsy itemdefinition before rendering a component.
Fixes the error case when in the codepanel, you type "-" indicating the beginning of a new component.

itemDef is then undefined and unwrapDefinntion with undefined will break the page

* removed testing code

* viewdef fixes

* type checking for monaco options

* first pass on error preview

* prototype for updating component def with parent of null

* Update sloterror.tsx

* cleanup and seperating PR's

* seperation of pr's

* cleanup

* Update workspace.json

* fixed watch

* wrapped around component renderer

* fix packing

* remove printlns

Co-authored-by: humandad <55447225+humandad@users.noreply.github.com>