feat: upgrade to react 19 (#4668)
For now, decision is to go with Option # (expose React* globals directly in ui project rather than via separate scripts).

If/When packui is removed and ui project is built directly with a bundler (e.g., esbuild), this decision can be revisited since it will be far simpler to use the bundler to generate the UMD react scripts since we'll be using a bundler anyway and would have to make multiple calls to it if we continue to use esbuild.

* feat: upgrade to react 19 WIP

* chore: format

* chore: improve types

* update package-lock

* fix types

* chore: add comment

* chore: complete todos

---------

Co-authored-by: humandad <55447225+humandad@users.noreply.github.com>